create GraphPartitioner interface, pull metis work out behind it, enable one to load it as a plugin
