+++ lib/Support/GraphPartitioner.cpp
@@ -1,0 +1,10 @@
+/*******************************************************************************
+ * Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                  *
+ * All rights reserved.                                                        *
+ *                                                                             *
+ * This source code and the accompanying materials are made available under    *
+ * the terms of the Apache License 2.0 which accompanies this distribution.    *
+ ******************************************************************************/
+#include "cudaq/Support/GraphPartioner.h"
+
+LLVM_INSTANTIATE_REGISTRY(cudaq::GraphPartitioner::RegistryType)
