+++ tools/cudaq-opt/cudaq-opt.cpp
@@ -47,8 +47,8 @@ static void registerInlinerExtension(mlir::DialectRegistry &registry) {
 
 /// @brief Add a command line flag for loading plugins
 static cl::list<std::string>
-    PassPlugins("load-pass-plugin",
-                cl::desc("Load passes from plugin library"));
+    CudaQPlugins("load-cudaq-plugin",
+                 cl::desc("Load CUDAQ plugin by specifying its library."));
 
 int main(int argc, char **argv) {
   // Set the bug report message to indicate users should file issues on
@@ -67,7 +67,7 @@ int main(int argc, char **argv) {
   // if so load it.
   std::vector<std::string> args(&argv[0], &argv[0] + argc);
   for (std::size_t i = 0; i < args.size(); i++) {
-    if (args[i].find("-load-pass-plugin") != std::string::npos) {
+    if (args[i].find("-load-cudaq-plugin") != std::string::npos) {
       auto Plugin = cudaq::Plugin::Load(args[i + 1]);
       if (!Plugin) {
         errs() << "Failed to load passes from '" << args[i + 1]
